Implementando regra para contabilizar os votos e dar o resultado da votação e refatorando o código.

+++ src/main/java/br/com/sicredi/votacao/service/PautaService.java
@@ -1,18 +1,18 @@
 package br.com.sicredi.votacao.service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.util.Assert;
 
+import br.com.sicredi.votacao.dto.PautaDTO;
+import br.com.sicredi.votacao.exception.PautaNotFoundException;
 import br.com.sicredi.votacao.model.Pauta;
 import br.com.sicredi.votacao.repository.PautaRepository;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 @Service
 public class PautaService {
 
@@ -20,45 +20,39 @@ public class PautaService {
 	private PautaRepository pautaRepository;
 	@Autowired
 	private VotoService votoService;
+	@Autowired
+	private SessaoService sessaoService;
+	@Autowired
+    private ModelMapper modelMapper;
 
-	public Pauta criar(Pauta pauta) {
-		Pauta pautaTest = pautaRepository.findByDescricaoPautaIgnoreCase(pauta.getDescricaoPauta());
-		Assert.isNull(pautaTest, "Pauta já cadastrada.");
+	public Pauta criar(PautaDTO pautaDTO) {
+		Pauta pauta = modelMapper.map(pautaDTO, Pauta.class);
 		return pautaRepository.save(pauta);
 	}
 
-	public Pauta recuperarPauta(Long id) {
+	public PautaDTO retornaPauta(Long id) {
 		Optional<Pauta> pautaOpt = pautaRepository.findById(id);
-		Assert.isTrue(pautaOpt.isPresent(), "Pauta não encontrada.");
-		return pautaOpt.get();
+		if (!pautaOpt.isPresent()) {
+			throw new PautaNotFoundException();
+		}
+		PautaDTO dto = modelMapper.map(pautaOpt.get(), PautaDTO.class);
+		return dto;
 	}
 
-	public List<Pauta> listarPautas() {
-		return pautaRepository.findAll();
+	public List<PautaDTO> listarPautas() {
+		return pautaRepository.findAll().stream().map(pauta -> modelMapper.map(pauta, PautaDTO.class))
+				.collect(Collectors.toList());
 	}
 	
 	public void delete(Long id) {
 		Optional<Pauta> pautaOpt = pautaRepository.findById(id);
-		Assert.isTrue(pautaOpt.isPresent(), "Pauta não encontrada.");
-        pautaRepository.delete(pautaOpt.get());
-        votoService.deleteByPautaId(id);
-    }
+		if (!pautaOpt.isPresent()) {
+			throw new PautaNotFoundException();
+		}
 
-	public void abrirPauta(Long idPauta, Integer qtdMinutos) {
-		log.debug("Inicio abertura de votacao pauta:" + idPauta);
-
-		Optional<Pauta> pautaOpt = pautaRepository.findById(idPauta);
-		Assert.isTrue(pautaOpt.isPresent(), "Pauta não encontrada.");
-
-		Pauta pauta = pautaOpt.get();
-
-		pauta.setDtVotacaoFim(LocalDateTime.now().plusMinutes(qtdMinutos));
-		pautaRepository.save(pauta);
-
-		votoService.deleteByPautaId(pauta.getId());
-
-		log.debug("Fim abertura de votacao pauta:" + idPauta);
-
+		pautaRepository.delete(pautaOpt.get());
+		sessaoService.deleteByPautaId(id);
+		votoService.deleteByPautaId(id);
 	}
 
 }
